Use the session's gpg-agent instance.
GnuPG >=2.1 enforces the use of one single gpg-agent instance. It is no
longer possible to start a separate one. passext now uses the gpg-agent
instance from the session.

We lose some of the control we had with our own gpg-agent instance, e.g.
that we're being asked for the passphrase at the start of the "session"
and the session's timeout.

On the other hand, the code is now a lot simpler, more stable and should
work with all different gpg versions.